Use `uv` tooling for build (#52)
This PR modifies the project to use `uv` instead of managing
dependencies with `requirements.txt` and `setup.py`. This move to more
modern tooling simplifies project onboarding, and adds a lock file for
the dependencies.

In builds associated with #49, tests are currently failing because of
mismatched dependency versions between my developer environment and the
CI system. The lock file should prevent this kind of issue and future
"works on my machine" situations.

This change also unifies the dependencies and build configuration into a
single file, instead of having 2 different requirement files for
different purposes, as well as a `setup.py` file which defines the build
process.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.