Disable check constraints via feature flag (#49)
This PR adds the `supports_table_check_constraints = False`
configuration, to skip any check constraints in migrated models.

Based on testing, I also defined an override for `_check_sql`, as some
code paths do not have an explicit check against
`supports_table_check_constraints` in the base class.

The tests validate that any check constraints in model definitions are
ignored, so as not to interrupt migrations with unsupported features.
The core model will still be created without the check constraints. This
was the intention of the previous code, but the empty queries defined by
the `schema.py` provided incomplete coverage.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.