Clean up expression index handling (#50)
This PR fixes up the definitions around expression indexes, which are
not supported by DSQL.

It defines the `supports_expression_indexes` feature flag to indicate
this, and adds tests to cover the wrapper-level behavior of the
`add_index` and `remove_index` overrides in the `schema.py` file.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.