Add documented support for Python 3.13 (#55)
This PR adds documented support for Python versions `3.13` and `3.14`.

`3.14` seems to be in prerelease so this might be getting ahead of
things a little bit, but if the workflows pass then this seems fine
since when it is released we won't need to make another change.

See [this
comment](https://github.com/awslabs/aurora-dsql-django/pull/52#discussion_r2379845044)
for original ask.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.