Add repo link to published metadata (#57)
This PR adds additional metadata so that PyPI can link back to this
repository. This will only take effect when this change is included in a
future release.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.