Switch to setuptools-scm for dynamic versioning (#58)
This PR uses
[setuptools-scm](https://setuptools-scm.readthedocs.io/en/latest/) to
avoid explicit version references in the source code.

This will make maintenance easier, as we don't need to worry about the
version in the source code becoming out of sync with GitHub releases.

In my testing, `uv build` uses the expected version number when a tag is
present, and uses a dev-specific version when no such tag is present. We
may additionally find this useful at some point if a customer reports an
issues showing a development version, as it means they are building from
modified source code which may be relevant.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.