Automatically update CHANGELOG.md on release (#59)
This PR adds a workflow which updates the `CHANGELOG.md` file
automatically based on the GitHub release notes.

We currently have release notes in 2 places, and they are not
synchronized. This will make it easier to maintain these and prevent
them from going out of sync.

When a GitHub release is created, we can automatically generate release
notes using the GitHub tools, and tweak them manually before publishing
the new version. This workflow will then pull the release notes using
the GitHub API, and create a PR which modifies the `CHANGELOG.md` file
for us to merge. I was originally going to have it push to the default
branch but dealing with conflicts if there were any would be a bit
weird. Having a PR should be sufficient and I don't expect there to be
any conflicts 99% of the time.

This is really difficult to test for a few reasons:
- It is intended to be triggered by a release, which we don't want to do
just to test
- I can't manually invoke the workflow until a version of the file is
available on the `version-0` (default) branch

I'd like to merge this as-is with minimal testing, and then make a
change on a dev branch to add a manual invocation, which can reference a
draft GitHub release which I won't publish. Until then the correctness
of this workflow is difficult to verify, but I will make sure to confirm
it works before publishing the next real release.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.