Remove no-op overrides for ALTER TABLE operations (#51)
This PR removes the current configuration for the following operations:

- `sql_create_pk`
- `sql_delete_constraint`
- `sql_delete_column`

Each of these operations uses an `ALTER TABLE` statement which is not
supported by DSQL. The previous approach was to replace the statement
with an empty query, but this allows the migration to continue despite
the operation having not been performed.

By removing the override, we fall back to the statement defined by the
base class which uses `ALTER TABLE`. This will result in the statement
being sent to DSQL, which will return an error. While this seemingly
causes migrations to fail where they previously didn't, the failure is
more correct in this scenario. By skipping the statement, we would be
leaving the database in a state which differs from what the migration
expects.

For customer migrations, these operations should be avoided. For Django
apps, manual steps may be needed to work around migrations that use
these operations.

It doesn't make a lot of sense to add a test for this, since we are now
falling back to the functionality already defined and tested in the base
class for the `schema.py` file.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.