Use built-in changelog PR creation (#61)
It turns out the `changelog-from-release` action supports creating the
PR directly, so we don't need the manual PR creation steps. I found this
because during testing it tried to push directly to the default branch
which was denied by our branch protection rules.

I've added a secondary checkout action since we will want to build and
publish the actual released tag, but then push `CHANGELOG.md` changes to
the default branch. This is a little confusing in the current state of
this workflow file since there are no build/publish steps, so I've added
a placeholder `TODO` to make it a bit clearer.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.