Fix foreign key references to UUID primary key type (#53)
This PR updates the configuration for `AutoField` and `BigAutoField` to
properly use the UUID type even when primary keys are reference through
foreign keys.

During testing, I found that running default app migrations would cause
an error like this:

```
django.db.utils.ProgrammingError: operator does not exist: integer = uuid
LINE 1: ...ent_type" ON ("auth_permission"."content_type_id" = CAST("dj..
```

After investigation, I found this was because the column was defined as
a foreign key, where the actual column being referenced was one of the
`AutoField` columns using a UUID. Since DSQL doesn't support foreign
keys, the type was dealt with internally in Django, which misunderstood
the type of the referenced column given it assumes `AutoField`s are
integers.

To correct this, I patched the `AutoField` and `BigAutoField` class in a
few places to return types which indicate they are using UUIDs. This
seems to fix the foreign key reference issue and prevent this error. I
checked the tables after running migrations and they both seem to have
the correct UUID type now.

I don't love that we need to directly patch the `AutoField` and
`BigAutoField` classes, but I couldn't find a way around it. I tried to
do the patching in the driver initializer so it is not susceptible to
module import variation, and should always apply correctly.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.