Disable server-side cursors by default (#54)
This PR configures a default value for `DISABLE_SERVER_SIDE_CURSORS`,
since server-side cursors are not supported by DSQL.

The default value here helps us avoid requiring customers to configure
this in their own project. It also makes it easier for us to enable
again if support is changed in the future.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.