Add publishing to release workflow (#63)
This PR modifies the `release.yml` workflow to publish the package to
PyPI.

It's difficult to test this without actually performing a release, so
there may be changes necessary after everything else is prepared for the
next release of the project.

The 2 publishing steps are done in environments, so we can add a manual
approval step to the production release environment. This will allow for
manual verification steps against the package published to the test
repository before the release is allowed continue.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.