Replace relative links in markdown files during build (#69)
This PR modifies the package build to replace relative markdown links
with absolute ones.

This is required for PyPI as the relative links do not work properly on
the release page. See [this
page](https://pypi.org/project/aurora-dsql-django/0.2.0/), which uses
`https://pypi.org/project/aurora-dsql-django/0.2.0/reference/KNOWN_ISSUES.md`
for the "Known Issues" link instead of a link to the GitHub repo.

I also added an additional step to the CI/CD workflow, since it was not
verifying the `uv build` for PRs. This is a good additional step to make
sure we don't break something in the release process. It is especially
relevant now since it will catch relative links that use `./` or `../`
as soon as they are introduced. We could technically support these kinds
of relative links but we don't use them right now and it's easier to
worry about that if/when we need them.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.