Add content type to PyPI description (#72)
This PR fixes the release build by correctly identifying the content
type of the description.

The previous release to PyPI failed since the content type was not
provided and PyPI could not determine it automatically.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.