Clone whole repo for tag metadata (#70)
This PR fixes a warning that occurs during the release build by cloning
the repo with all history.

```
/home/runner/work/_temp/setup-uv-cache/.../lib/python3.12/site-packages/setuptools_scm/git.py:202:
UserWarning: "/home/runner/work/aurora-dsql-django/aurora-dsql-django" is shallow and may cause errors
```

I think the warning is harmless in practice since the tag information is
able to be retrieved based on the successful release, but this should
prevent the warning and any future potential issues.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.