Remove Python 3.8 support (#74)
This PR removes Python 3.8, since it is end of life as of October 2024
according to [this page](https://devguide.python.org/versions/). We may
need to do the same for 3.9 soon since it seems to be end of life as of
October 2025.

The main reason for removing 3.8 support here is to enable upgrade of
`setuptools`. This should help to fix a build warning which will become
an issue as of February 2026.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.